back-end/requirements.txt
flask
flask-cors
openai
python-dotenv
sqlalchemy
biopython
feedparser
gunicorn